+++ src/query_2pcf_module.F90
@@ -61,7 +61,8 @@ contains
     unit_num = 30
     print *, 'writing output to: ', trim(mode_output_file('2pcf'))
     open(unit_num, file=trim(mode_output_file('2pcf')), status='unknown')
-    write(unit_num, *) 'r min, r max, mu min, mu max, NN, RR, 2pcf (xi)'
+    call write_provenance(unit_num)
+    write(unit_num, '(a)') '# r min, r max, mu min, mu max, NN, RR, 2pcf (xi)'
 
     do l = 1, cfg%nbins
       do k = 1, cfg%nmu